requirements.txt
Flask
Werkzeug
requests
ffmpeg-python
openai-whisper
gunicorn
APScheduler
srt
numpy
torch
torchaudio
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-api-core
google-cloud-storage
google-cloud-run
psutil
boto3
Pillow
matplotlib
yt-dlp
chatterbox-tts
flasgger